+++ app/grouping/create/components/inputSection/InputArea.server.tsx
@@ -1,14 +1,14 @@
-import type { InputHTMLAttributes } from 'react';
-
 import { Input } from '@/components/common/Input';
 import { Spacing } from '@/components/common/Spacing';
 
-interface InputSectionProps extends InputHTMLAttributes<HTMLInputElement> {
+import type { InputHTMLAttributes } from 'react';
+
+interface InputAreaProps extends InputHTMLAttributes<HTMLInputElement> {
   title: string;
   onClick: () => void;
 }
 
-export default function InputSection({ title, onClick, ...props }: InputSectionProps) {
+export default function InputArea({ title, onClick, ...props }: InputAreaProps) {
   return (
     <section onClick={onClick}>
       <p className="text-14">{title}</p>
